All zero values too need to be poll and written together with date in a DB.
It is necessary to do check on NULL, but not on zero (0) values when polling. Otherwise, if all values zero (0), in a DB it isn't write information on the current polling. There can be a situation when in cabnavigator or when viewing CDU tooltip the last __nonzero__ value for __some__ __long__ __ago__ passed date is displayed :-/. I.e., for example, the current consumption of CDU is zero, and is displayed wrong nonzero value for some long ago passed date.

Result: in all type MeasurePoint have to be poll and write to the DB with the current date and time, isn't dependent on that zero values or not.